Move process_device_log to phonology
