+++ corehq/ex-submodules/couchforms/util.py
@@ -17,13 +17,11 @@ from corehq.apps.tzmigration import timezone_migration_in_progress
 from corehq.form_processor.utils import new_xform, acquire_lock_for_xform
 from corehq.form_processor.interfaces.processor import FormProcessorInterface
 from corehq.util.soft_assert import soft_assert
-from corehq.apps.users.dbaccessors import get_user_id_by_username
-from corehq.apps.users.util import format_username
-from phonelog.models import UserEntry, DeviceReportEntry
 from dimagi.utils.couch.undo import DELETED_SUFFIX
 from dimagi.utils.logging import notify_exception
 from dimagi.utils.mixins import UnicodeMixIn
 from dimagi.utils.couch import LockManager
+from phonelog.utils import process_device_log
 import couchforms
 from .const import BadRequest, DEVICE_LOG_XMLNS
 from .exceptions import DuplicateError, UnexpectedDeletedXForm, \
@@ -537,65 +535,3 @@ def spoof_submission(submit_url, body, name="form.xml", hqsubmission=True,
         xform['doc_type'] = "HQSubmission"
         xform.save()
     return response
-
-
-def _force_list(obj_or_list):
-    return obj_or_list if isinstance(obj_or_list, list) else [obj_or_list]
-
-
-def _get_logs(form, report_name, report_slug):
-    report = form.get(report_name, {}) or {}
-    if isinstance(report, list):
-        return filter(None, [log.get(report_slug) for log in report])
-    return report.get(report_slug, [])
-
-
-def process_device_log(domain, xform):
-    form_data = xform.form_data
-    userlogs = _get_logs(form_data, 'user_subreport', 'user')
-    UserEntry.objects.filter(xform_id=xform.form_id).delete()
-    DeviceReportEntry.objects.filter(xform_id=xform.form_id).delete()
-    to_save = []
-    for i, log in enumerate(_force_list(userlogs)):
-        to_save.append(UserEntry(
-            xform_id=xform.form_id,
-            i=i,
-            user_id=log["user_id"],
-            username=log["username"],
-            sync_token=log["sync_token"],
-        ))
-    UserEntry.objects.bulk_create(to_save)
-
-    logs = _get_logs(form_data, 'log_subreport', 'log')
-    logged_in_username = None
-    logged_in_user_id = None
-    to_save = []
-    for i, log in enumerate(_force_list(logs)):
-        if not log:
-            continue
-        if log["type"] == 'login':
-            # j2me log = user_id_prefix-username
-            logged_in_username = log["msg"].split('-')[1]
-            cc_username = format_username(logged_in_username, domain)
-            logged_in_user_id = get_user_id_by_username(cc_username)
-        elif log["type"] == 'user' and log["msg"][:5] == 'login':
-            # android log = login|username|user_id
-            msg_split = log["msg"].split('|')
-            logged_in_username = msg_split[1]
-            logged_in_user_id = msg_split[2]
-
-        to_save.append(DeviceReportEntry(
-            xform_id=xform.form_id,
-            i=i,
-            domain=domain,
-            type=log["type"],
-            msg=log["msg"],
-            # must accept either date or datetime string
-            date=log["@date"],
-            server_date=xform.received_on,
-            app_version=form_data.get('app_version'),
-            device_id=form_data.get('device_id'),
-            username=logged_in_username,
-            user_id=logged_in_user_id,
-        ))
-    DeviceReportEntry.objects.bulk_create(to_save)
